ops-coffee-blog/requirements.txt
Django==2.2
jinja2==2.10
markdown==3.1.1
gitpython==2.1.11